Cleanup formatter module
Remove the additional module used for examples generation.
